+++ src/main/java/subway/domain/line/controller/PathController.java
@@ -1,4 +1,4 @@
-package subway.domain.path.controller;
+package subway.domain.line.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import subway.domain.path.domain.Path;
-import subway.domain.path.dto.PathResponse;
-import subway.domain.path.service.PathService;
+import subway.domain.line.domain.Path;
+import subway.domain.line.dto.PathResponse;
+import subway.domain.line.service.PathService;
 import subway.global.common.ResultResponse;
 
 @Controller
