+++ src/main/java/subway/domain/line/domain/FareCalculator.java
@@ -1,4 +1,4 @@
-package subway.domain.path.domain;
+package subway.domain.line.domain;
 
 public enum FareCalculator {
