+++ src/test/java/subway/path/service/PathServiceTest.java
@@ -8,8 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.jdbc.Sql;
-import subway.domain.path.domain.Path;
-import subway.domain.path.service.PathService;
+import subway.domain.line.domain.Path;
+import subway.domain.line.service.PathService;
 import subway.domain.line.entity.StationEntity;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
